+++ internal/provider/data_source_nxos_port_channel_interface.go
@@ -57,7 +57,7 @@ func (d *PortChannelInterfaceDataSource) Metadata(_ context.Context, req datasou
 func (d *PortChannelInterfaceDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This data source can read the configuration of port-channel interfaces on NX-OS devices, including channel mode, member link settings, switchport mode, and VLAN assignments.").AddApiDocumentation("interfaceEntity", "", []string{"pcAggrIf", "nwRtVrfMbr", "pcRsMbrIfs"}, []string{"Interfaces/pc:AggrIf/", "Routing%20and%20Forwarding/nw:RtVrfMbr/", "Interfaces/pc:RsMbrIfs/"}).String,
+		MarkdownDescription: helpers.NewResourceDescription("This data source can read the configuration of port-channel interfaces on NX-OS devices, including channel mode, member link settings, switchport mode, and VLAN assignments.").AddApiDocumentation("interfaceEntity", "", []string{"pcAggrIf", "nwRtVrfMbr", "l1StormCtrlP", "pcRsMbrIfs"}, []string{"Interfaces/pc:AggrIf/", "Routing%20and%20Forwarding/nw:RtVrfMbr/", "System/l1:StormCtrlP/", "Interfaces/pc:RsMbrIfs/"}).String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -241,6 +241,26 @@ func (d *PortChannelInterfaceDataSource) Schema(ctx context.Context, req datasou
 							MarkdownDescription: "DN of VRF. For example: `sys/inst-VRF1`.",
 							Computed:            true,
 						},
+						"storm_control_burst_packets_per_second": schema.Int64Attribute{
+							MarkdownDescription: "Max burst size.",
+							Computed:            true,
+						},
+						"storm_control_burst_rate": schema.StringAttribute{
+							MarkdownDescription: "Max burst size.",
+							Computed:            true,
+						},
+						"storm_control_rate": schema.StringAttribute{
+							MarkdownDescription: "Traffic rate.",
+							Computed:            true,
+						},
+						"storm_control_rate_packets_per_second": schema.Int64Attribute{
+							MarkdownDescription: "Tarffic rate.",
+							Computed:            true,
+						},
+						"storm_control_packet_type": schema.StringAttribute{
+							MarkdownDescription: "Packet Type.",
+							Computed:            true,
+						},
 						"members": schema.MapNestedAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("List of port-channel member interfaces.\n  - Map key: `interface_dn` - DN of interface. For example: `sys/intf/phys-[eth1/1]`.").String,
 							Computed:            true,
@@ -288,7 +308,7 @@ func (d *PortChannelInterfaceDataSource) Read(ctx context.Context, req datasourc
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to find device '%s' in provider configuration", config.Device.ValueString()))
 		return
 	}
-	queries := []func(*nxos.Req){nxos.Query("rsp-subtree", "full"), nxos.Query("rsp-subtree-class", "pcAggrIf,nwRtVrfMbr,pcRsMbrIfs")}
+	queries := []func(*nxos.Req){nxos.Query("rsp-subtree", "full"), nxos.Query("rsp-subtree-class", "pcAggrIf,nwRtVrfMbr,l1StormCtrlP,pcRsMbrIfs")}
 	res, err := device.Client.GetDn(config.getDn(), queries...)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))