+++ internal/provider/model_nxos_physical_interface.go
@@ -43,61 +43,66 @@ type PhysicalInterface struct {
 }
 
 type PhysicalInterfacePhysicalInterfaces struct {
-	FecMode                        types.String `tfsdk:"fec_mode"`
-	AccessVlan                     types.String `tfsdk:"access_vlan"`
-	AdminState                     types.String `tfsdk:"admin_state"`
-	AutoNegotiation                types.String `tfsdk:"auto_negotiation"`
-	Bandwidth                      types.Int64  `tfsdk:"bandwidth"`
-	Delay                          types.Int64  `tfsdk:"delay"`
-	Description                    types.String `tfsdk:"description"`
-	Duplex                         types.String `tfsdk:"duplex"`
-	Layer                          types.String `tfsdk:"layer"`
-	LinkLogging                    types.String `tfsdk:"link_logging"`
-	LinkDebounceDown               types.Int64  `tfsdk:"link_debounce_down"`
-	LinkDebounceUp                 types.Int64  `tfsdk:"link_debounce_up"`
-	Medium                         types.String `tfsdk:"medium"`
-	Mode                           types.String `tfsdk:"mode"`
-	Mtu                            types.Int64  `tfsdk:"mtu"`
-	NativeVlan                     types.String `tfsdk:"native_vlan"`
-	Speed                          types.String `tfsdk:"speed"`
-	SpeedGroup                     types.String `tfsdk:"speed_group"`
-	TrunkVlans                     types.String `tfsdk:"trunk_vlans"`
-	UniDirectionalEthernet         types.String `tfsdk:"uni_directional_ethernet"`
-	UserConfiguredFlags            types.String `tfsdk:"user_configured_flags"`
-	Beacon                         types.String `tfsdk:"beacon"`
-	DfeAdaptiveTuning              types.String `tfsdk:"dfe_adaptive_tuning"`
-	DfeTuningDelay                 types.Int64  `tfsdk:"dfe_tuning_delay"`
-	Dot1qEtherType                 types.Int64  `tfsdk:"dot1q_ether_type"`
-	EqualizationDelay              types.Int64  `tfsdk:"equalization_delay"`
-	InheritBandwidth               types.Int64  `tfsdk:"inherit_bandwidth"`
-	ItuChannel                     types.Int64  `tfsdk:"itu_channel"`
-	LinkActiveJitterManagement     types.String `tfsdk:"link_active_jitter_management"`
-	LinkFlapErrorDisable           types.String `tfsdk:"link_flap_error_disable"`
-	LinkFlapErrorMax               types.Int64  `tfsdk:"link_flap_error_max"`
-	LinkFlapErrorSeconds           types.Int64  `tfsdk:"link_flap_error_seconds"`
-	LinkLoopback                   types.String `tfsdk:"link_loopback"`
-	LinkMacUpTimer                 types.Int64  `tfsdk:"link_mac_up_timer"`
-	LinkMaxBringUpTimer            types.Int64  `tfsdk:"link_max_bring_up_timer"`
-	LinkTransmitReset              types.String `tfsdk:"link_transmit_reset"`
-	Mdix                           types.String `tfsdk:"mdix"`
-	MediaType                      types.String `tfsdk:"media_type"`
-	OpticsLoopback                 types.String `tfsdk:"optics_loopback"`
-	PacketTimestampEgressSourceId  types.Int64  `tfsdk:"packet_timestamp_egress_source_id"`
-	PacketTimestampIngressSourceId types.Int64  `tfsdk:"packet_timestamp_ingress_source_id"`
-	PacketTimestampState           types.String `tfsdk:"packet_timestamp_state"`
-	PortType                       types.String `tfsdk:"port_type"`
-	RouterMac                      types.String `tfsdk:"router_mac"`
-	SnmpTrapState                  types.String `tfsdk:"snmp_trap_state"`
-	SpanMode                       types.String `tfsdk:"span_mode"`
-	Squelch                        types.String `tfsdk:"squelch"`
-	TransparentMode                types.String `tfsdk:"transparent_mode"`
-	TrunkLogging                   types.String `tfsdk:"trunk_logging"`
-	Usage                          types.String `tfsdk:"usage"`
-	VoicePortCos                   types.Int64  `tfsdk:"voice_port_cos"`
-	VoicePortTrust                 types.String `tfsdk:"voice_port_trust"`
-	VoiceVlanId                    types.Int64  `tfsdk:"voice_vlan_id"`
-	VoiceVlanType                  types.String `tfsdk:"voice_vlan_type"`
-	VrfDn                          types.String `tfsdk:"vrf_dn"`
+	FecMode                           types.String `tfsdk:"fec_mode"`
+	AccessVlan                        types.String `tfsdk:"access_vlan"`
+	AdminState                        types.String `tfsdk:"admin_state"`
+	AutoNegotiation                   types.String `tfsdk:"auto_negotiation"`
+	Bandwidth                         types.Int64  `tfsdk:"bandwidth"`
+	Delay                             types.Int64  `tfsdk:"delay"`
+	Description                       types.String `tfsdk:"description"`
+	Duplex                            types.String `tfsdk:"duplex"`
+	Layer                             types.String `tfsdk:"layer"`
+	LinkLogging                       types.String `tfsdk:"link_logging"`
+	LinkDebounceDown                  types.Int64  `tfsdk:"link_debounce_down"`
+	LinkDebounceUp                    types.Int64  `tfsdk:"link_debounce_up"`
+	Medium                            types.String `tfsdk:"medium"`
+	Mode                              types.String `tfsdk:"mode"`
+	Mtu                               types.Int64  `tfsdk:"mtu"`
+	NativeVlan                        types.String `tfsdk:"native_vlan"`
+	Speed                             types.String `tfsdk:"speed"`
+	SpeedGroup                        types.String `tfsdk:"speed_group"`
+	TrunkVlans                        types.String `tfsdk:"trunk_vlans"`
+	UniDirectionalEthernet            types.String `tfsdk:"uni_directional_ethernet"`
+	UserConfiguredFlags               types.String `tfsdk:"user_configured_flags"`
+	Beacon                            types.String `tfsdk:"beacon"`
+	DfeAdaptiveTuning                 types.String `tfsdk:"dfe_adaptive_tuning"`
+	DfeTuningDelay                    types.Int64  `tfsdk:"dfe_tuning_delay"`
+	Dot1qEtherType                    types.Int64  `tfsdk:"dot1q_ether_type"`
+	EqualizationDelay                 types.Int64  `tfsdk:"equalization_delay"`
+	InheritBandwidth                  types.Int64  `tfsdk:"inherit_bandwidth"`
+	ItuChannel                        types.Int64  `tfsdk:"itu_channel"`
+	LinkActiveJitterManagement        types.String `tfsdk:"link_active_jitter_management"`
+	LinkFlapErrorDisable              types.String `tfsdk:"link_flap_error_disable"`
+	LinkFlapErrorMax                  types.Int64  `tfsdk:"link_flap_error_max"`
+	LinkFlapErrorSeconds              types.Int64  `tfsdk:"link_flap_error_seconds"`
+	LinkLoopback                      types.String `tfsdk:"link_loopback"`
+	LinkMacUpTimer                    types.Int64  `tfsdk:"link_mac_up_timer"`
+	LinkMaxBringUpTimer               types.Int64  `tfsdk:"link_max_bring_up_timer"`
+	LinkTransmitReset                 types.String `tfsdk:"link_transmit_reset"`
+	Mdix                              types.String `tfsdk:"mdix"`
+	MediaType                         types.String `tfsdk:"media_type"`
+	OpticsLoopback                    types.String `tfsdk:"optics_loopback"`
+	PacketTimestampEgressSourceId     types.Int64  `tfsdk:"packet_timestamp_egress_source_id"`
+	PacketTimestampIngressSourceId    types.Int64  `tfsdk:"packet_timestamp_ingress_source_id"`
+	PacketTimestampState              types.String `tfsdk:"packet_timestamp_state"`
+	PortType                          types.String `tfsdk:"port_type"`
+	RouterMac                         types.String `tfsdk:"router_mac"`
+	SnmpTrapState                     types.String `tfsdk:"snmp_trap_state"`
+	SpanMode                          types.String `tfsdk:"span_mode"`
+	Squelch                           types.String `tfsdk:"squelch"`
+	TransparentMode                   types.String `tfsdk:"transparent_mode"`
+	TrunkLogging                      types.String `tfsdk:"trunk_logging"`
+	Usage                             types.String `tfsdk:"usage"`
+	VoicePortCos                      types.Int64  `tfsdk:"voice_port_cos"`
+	VoicePortTrust                    types.String `tfsdk:"voice_port_trust"`
+	VoiceVlanId                       types.Int64  `tfsdk:"voice_vlan_id"`
+	VoiceVlanType                     types.String `tfsdk:"voice_vlan_type"`
+	VrfDn                             types.String `tfsdk:"vrf_dn"`
+	StormControlBurstPacketsPerSecond types.Int64  `tfsdk:"storm_control_burst_packets_per_second"`
+	StormControlBurstRate             types.String `tfsdk:"storm_control_burst_rate"`
+	StormControlRate                  types.String `tfsdk:"storm_control_rate"`
+	StormControlRatePacketsPerSecond  types.Int64  `tfsdk:"storm_control_rate_packets_per_second"`
+	StormControlPacketType            types.String `tfsdk:"storm_control_packet_type"`
 }
 
 type PhysicalInterfaceIdentity struct {
@@ -321,6 +326,25 @@ func (data PhysicalInterface) toBody(config PhysicalInterface) nxos.Body {
 			if attrs != "{}" {
 				body, _ = sjson.SetRaw(body, nestedChildrenPath+".-1.nwRtVrfMbr.attributes", attrs)
 			}
+			attrs = "{}"
+			if !child.StormControlBurstPacketsPerSecond.IsUnknown() && !child.StormControlBurstPacketsPerSecond.IsNull() {
+				attrs, _ = sjson.Set(attrs, "burstPps", strconv.FormatInt(child.StormControlBurstPacketsPerSecond.ValueInt64(), 10))
+			}
+			if !child.StormControlBurstRate.IsUnknown() && !child.StormControlBurstRate.IsNull() {
+				attrs, _ = sjson.Set(attrs, "burstRate", child.StormControlBurstRate.ValueString())
+			}
+			if !child.StormControlRate.IsUnknown() && !child.StormControlRate.IsNull() {
+				attrs, _ = sjson.Set(attrs, "rate", child.StormControlRate.ValueString())
+			}
+			if !child.StormControlRatePacketsPerSecond.IsUnknown() && !child.StormControlRatePacketsPerSecond.IsNull() {
+				attrs, _ = sjson.Set(attrs, "ratePps", strconv.FormatInt(child.StormControlRatePacketsPerSecond.ValueInt64(), 10))
+			}
+			if !child.StormControlPacketType.IsUnknown() && !child.StormControlPacketType.IsNull() {
+				attrs, _ = sjson.Set(attrs, "type", child.StormControlPacketType.ValueString())
+			}
+			if attrs != "{}" {
+				body, _ = sjson.SetRaw(body, nestedChildrenPath+".-1.l1StormCtrlP.attributes", attrs)
+			}
 		}
 	}
 
@@ -406,6 +430,24 @@ func (data *PhysicalInterface) fromBody(res gjson.Result) {
 								},
 							)
 							child.VrfDn = types.StringValue(rnwRtVrfMbr.Get("nwRtVrfMbr.attributes.tDn").String())
+						}
+						{
+							var rl1StormCtrlP gjson.Result
+							value.Get("children").ForEach(
+								func(_, nestedV gjson.Result) bool {
+									rnValue := nestedV.Get("l1StormCtrlP.attributes.rn").String()
+									if rnValue == "stormctrlp" {
+										rl1StormCtrlP = nestedV
+										return false
+									}
+									return true
+								},
+							)
+							child.StormControlBurstPacketsPerSecond = types.Int64Value(rl1StormCtrlP.Get("l1StormCtrlP.attributes.burstPps").Int())
+							child.StormControlBurstRate = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.burstRate").String())
+							child.StormControlRate = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.rate").String())
+							child.StormControlRatePacketsPerSecond = types.Int64Value(rl1StormCtrlP.Get("l1StormCtrlP.attributes.ratePps").Int())
+							child.StormControlPacketType = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.type").String())
 						}
 						if data.PhysicalInterfaces == nil {
 							data.PhysicalInterfaces = make(map[string]PhysicalInterfacePhysicalInterfaces)
@@ -726,6 +768,44 @@ func (data *PhysicalInterface) updateFromBody(res gjson.Result) {
 				item.VrfDn = types.StringValue(rnwRtVrfMbr.Get("nwRtVrfMbr.attributes.tDn").String())
 			} else {
 				item.VrfDn = types.StringNull()
+			}
+		}
+		{
+			var rl1StormCtrlP gjson.Result
+			rl1PhysIf.Get("l1PhysIf.children").ForEach(
+				func(_, v gjson.Result) bool {
+					rnValue := v.Get("l1StormCtrlP.attributes.rn").String()
+					if rnValue == "stormctrlp" {
+						rl1StormCtrlP = v
+						return false
+					}
+					return true
+				},
+			)
+			if !item.StormControlBurstPacketsPerSecond.IsNull() {
+				item.StormControlBurstPacketsPerSecond = types.Int64Value(rl1StormCtrlP.Get("l1StormCtrlP.attributes.burstPps").Int())
+			} else {
+				item.StormControlBurstPacketsPerSecond = types.Int64Null()
+			}
+			if !item.StormControlBurstRate.IsNull() {
+				item.StormControlBurstRate = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.burstRate").String())
+			} else {
+				item.StormControlBurstRate = types.StringNull()
+			}
+			if !item.StormControlRate.IsNull() {
+				item.StormControlRate = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.rate").String())
+			} else {
+				item.StormControlRate = types.StringNull()
+			}
+			if !item.StormControlRatePacketsPerSecond.IsNull() {
+				item.StormControlRatePacketsPerSecond = types.Int64Value(rl1StormCtrlP.Get("l1StormCtrlP.attributes.ratePps").Int())
+			} else {
+				item.StormControlRatePacketsPerSecond = types.Int64Null()
+			}
+			if !item.StormControlPacketType.IsNull() {
+				item.StormControlPacketType = types.StringValue(rl1StormCtrlP.Get("l1StormCtrlP.attributes.type").String())
+			} else {
+				item.StormControlPacketType = types.StringNull()
 			}
 		}
 		data.PhysicalInterfaces[key] = item