+++ internal/provider/resource_nxos_port_channel_interface.go
@@ -63,7 +63,7 @@ func (r *PortChannelInterfaceResource) Metadata(ctx context.Context, req resourc
 func (r *PortChannelInterfaceResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This resource can manage port-channel interfaces on NX-OS devices, including channel mode, member link settings, switchport mode, and VLAN assignments.").AddApiDocumentation("interfaceEntity", "", []string{"pcAggrIf", "nwRtVrfMbr", "pcRsMbrIfs"}, []string{"Interfaces/pc:AggrIf/", "Routing%20and%20Forwarding/nw:RtVrfMbr/", "Interfaces/pc:RsMbrIfs/"}).String,
+		MarkdownDescription: helpers.NewResourceDescription("This resource can manage port-channel interfaces on NX-OS devices, including channel mode, member link settings, switchport mode, and VLAN assignments.").AddApiDocumentation("interfaceEntity", "", []string{"pcAggrIf", "nwRtVrfMbr", "l1StormCtrlP", "pcRsMbrIfs"}, []string{"Interfaces/pc:AggrIf/", "Routing%20and%20Forwarding/nw:RtVrfMbr/", "System/l1:StormCtrlP/", "Interfaces/pc:RsMbrIfs/"}).String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -352,6 +352,35 @@ func (r *PortChannelInterfaceResource) Schema(ctx context.Context, req resource.
 							MarkdownDescription: helpers.NewAttributeDescription("DN of VRF. For example: `sys/inst-VRF1`.").String,
 							Optional:            true,
 						},
+						"storm_control_burst_packets_per_second": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Max burst size.").AddIntegerRangeDescription(0, 4294967295).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(0, 4294967295),
+							},
+						},
+						"storm_control_burst_rate": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Max burst size.").String,
+							Optional:            true,
+						},
+						"storm_control_rate": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Traffic rate.").String,
+							Optional:            true,
+						},
+						"storm_control_rate_packets_per_second": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Tarffic rate.").AddIntegerRangeDescription(0, 4294967295).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(0, 4294967295),
+							},
+						},
+						"storm_control_packet_type": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Packet Type.").AddStringEnumDescription("bcast", "unk-ucast", "mcast", "all").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("bcast", "unk-ucast", "mcast", "all"),
+							},
+						},
 						"members": schema.MapNestedAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("List of port-channel member interfaces.\n  - Map key: `interface_dn` - DN of interface. For example: `sys/intf/phys-[eth1/1]`.").String,
 							Optional:            true,
@@ -476,7 +505,7 @@ func (r *PortChannelInterfaceResource) Read(ctx context.Context, req resource.Re
 	}
 
 	if device.Managed {
-		queries := []func(*nxos.Req){nxos.Query("rsp-subtree", "full"), nxos.Query("rsp-subtree-class", "pcAggrIf,nwRtVrfMbr,pcRsMbrIfs")}
+		queries := []func(*nxos.Req){nxos.Query("rsp-subtree", "full"), nxos.Query("rsp-subtree-class", "pcAggrIf,nwRtVrfMbr,l1StormCtrlP,pcRsMbrIfs")}
 		res, err := device.Client.GetDn(state.Dn.ValueString(), queries...)
 		if err != nil {
 			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))